Generate an index for WritingPractise feature
